Podcast: consolidate Posts to Podcast feature into its own subdirectory
