+++ projects/packages/podcast/src/posts-to-podcast/class-admin-page.php
@@ -7,36 +7,22 @@
  * posts picker, submits the generate request, polls the job, and resumes
  * across reloads via localStorage.
  *
- * Bootstrapped from `Podcast::init()` after the Host (Simple/WoA) and
- * `jetpack_podcast_untangle` gates have already been checked upstream.
- *
  * @package automattic/jetpack-podcast
  */
 
-namespace Automattic\Jetpack\Podcast;
+namespace Automattic\Jetpack\Podcast\Posts_To_Podcast;
 
-require_once __DIR__ . '/admin-pages/create-ai-podcast/presets.php';
-
-use Automattic\Jetpack\Assets;
-use Automattic\Jetpack\Status\Host;
-use function Automattic\Jetpack\Podcast\Admin_Pages\Create_AI_Podcast\length_presets;
-use function Automattic\Jetpack\Podcast\Admin_Pages\Create_AI_Podcast\voice_presets;
-use function Automattic\Jetpack\Podcast\Admin_Pages\Create_AI_Podcast\window_presets;
+require_once __DIR__ . '/admin-page/presets.php';
 
 /**
  * Registers the Media > Create AI Podcast submenu and renders the page.
  */
-class Create_AI_Podcast_Page {
+class Admin_Page {
 
 	const PAGE_SLUG         = 'create-ai-podcast';
 	const SCRIPT_HANDLE     = 'jetpack-create-ai-podcast';
 	const STYLE_HANDLE      = 'jetpack-create-ai-podcast';
 	const EPISODES_PER_PAGE = 5;
-
-	const POST_PUBLISH_PROMO_SCRIPT_HANDLE    = 'jetpack-post-publish-podcast-promo';
-	const POST_PUBLISH_PROMO_DISMISSED_OPTION = 'jetpack_posts_to_podcast_post_publish_promo_dismissed';
-	const POST_PUBLISH_PROMO_MIN_POSTS        = 5;
-	const POST_PUBLISH_PROMO_MIN_VISITORS     = 50;
 
 	/**
 	 * Whether `init()` has wired its hooks.
@@ -55,7 +41,6 @@ class Create_AI_Podcast_Page {
 		self::$initialized = true;
 
 		add_action( 'admin_menu', array( __CLASS__, 'register_menu' ) );
-		add_action( 'enqueue_block_editor_assets', array( __CLASS__, 'enqueue_post_publish_promo_assets' ) );
 	}
 
 	/**
@@ -101,8 +86,8 @@ class Create_AI_Podcast_Page {
 	 * Enqueue the static JS island, the page stylesheet, and the localized data bundle.
 	 */
 	public static function enqueue_assets() {
-		$base_url  = plugins_url( 'admin-pages/create-ai-podcast/', __FILE__ );
-		$base_path = __DIR__ . '/admin-pages/create-ai-podcast/';
+		$base_url  = plugins_url( 'admin-page/', __FILE__ );
+		$base_path = __DIR__ . '/admin-page/';
 
 		wp_enqueue_style(
 			self::STYLE_HANDLE,
@@ -124,216 +109,6 @@ class Create_AI_Podcast_Page {
 			'jetpackCreateAiPodcast',
 			self::build_localized_data()
 		);
-	}
-
-	/**
-	 * Enqueue the post-publish modal in the post block editor for eligible sites.
-	 */
-	public static function enqueue_post_publish_promo_assets() {
-		if (
-			! self::is_post_block_editor()
-			|| self::is_current_post_published_for_post_publish_promo()
-			|| ! self::is_post_publish_promo_site_eligible()
-		) {
-			return;
-		}
-
-		Assets::register_script(
-			self::POST_PUBLISH_PROMO_SCRIPT_HANDLE,
-			'../dist/blocks/post-publish-podcast-promo/editor.js',
-			__FILE__,
-			array(
-				'enqueue'    => true,
-				'in_footer'  => true,
-				'textdomain' => 'jetpack-podcast',
-			)
-		);
-
-		wp_add_inline_script(
-			self::POST_PUBLISH_PROMO_SCRIPT_HANDLE,
-			'window.jetpackPostPublishPodcastPromo = ' . wp_json_encode(
-				array(
-					'createUrl'   => admin_url( 'upload.php?page=' . self::PAGE_SLUG ),
-					'dismissPath' => Posts_To_Podcast_Endpoint::get_post_publish_promo_dismiss_rest_path(),
-				),
-				JSON_HEX_TAG | JSON_HEX_AMP | JSON_HEX_APOS | JSON_HEX_QUOT
-			) . ';',
-			'before'
-		);
-	}
-
-	/**
-	 * Whether the current editor post has already been published.
-	 */
-	private static function is_current_post_published_for_post_publish_promo(): bool {
-		$post = get_post();
-
-		return $post instanceof \WP_Post
-			&& 'post' === $post->post_type
-			&& 'publish' === $post->post_status;
-	}
-
-	/**
-	 * Whether the current screen is the post block editor.
-	 */
-	private static function is_post_block_editor(): bool {
-		if ( ! function_exists( 'get_current_screen' ) ) {
-			return false;
-		}
-
-		$screen = get_current_screen();
-		return ! empty( $screen )
-			&& 'post' === $screen->base
-			&& 'post' === $screen->post_type
-			&& $screen->is_block_editor();
-	}
-
-	/**
-	 * Whether the site has enough published posts to generate a better episode.
-	 */
-	private static function has_enough_recent_posts_for_post_publish_promo(): bool {
-		/**
-		 * Filters the minimum posts published in the last month needed for the Posts to Podcast post-publish promo.
-		 *
-		 * @since 1.0.0
-		 *
-		 * @param int $minimum Minimum number of published posts.
-		 */
-		$minimum = (int) apply_filters(
-			'jetpack_posts_to_podcast_post_publish_promo_min_published_posts',
-			self::POST_PUBLISH_PROMO_MIN_POSTS
-		);
-		$minimum = max( 1, $minimum );
-
-		$published_posts = get_posts(
-			array(
-				'fields'           => 'ids',
-				'no_found_rows'    => true,
-				'post_status'      => 'publish',
-				'post_type'        => 'post',
-				'posts_per_page'   => $minimum,
-				'suppress_filters' => false,
-				'date_query'       => array(
-					array(
-						'after'     => '1 month ago',
-						'inclusive' => true,
-					),
-				),
-			)
-		);
-		$total           = count( $published_posts );
-
-		$post = get_post();
-		if ( $post && 'post' === $post->post_type && 'publish' !== $post->post_status ) {
-			++$total;
-		}
-
-		return $total >= $minimum;
-	}
-
-	/**
-	 * Whether the site has visitors who could benefit from a podcast episode.
-	 */
-	private static function has_visitors_for_post_publish_promo(): bool {
-		$visitors = self::get_post_publish_promo_visitor_count();
-
-		/**
-		 * Filters the minimum visitors in the last week needed for the Posts to Podcast post-publish promo.
-		 *
-		 * @since 1.0.0
-		 *
-		 * @param int $minimum Minimum number of visitors.
-		 */
-		$minimum = (int) apply_filters(
-			'jetpack_posts_to_podcast_post_publish_promo_min_visitors',
-			self::POST_PUBLISH_PROMO_MIN_VISITORS
-		);
-
-		return $visitors >= max( 1, $minimum );
-	}
-
-	/**
-	 * Fetch the last week's visitor count from Jetpack Stats when available.
-	 */
-	private static function get_post_publish_promo_visitor_count(): int {
-		$host = new Host();
-		if ( $host->is_wpcom_simple() ) {
-			return self::get_wpcom_simple_post_publish_promo_visitor_count();
-		}
-
-		if ( class_exists( '\Automattic\Jetpack\Stats\WPCOM_Stats' ) ) {
-			$wpcom_stats = new \Automattic\Jetpack\Stats\WPCOM_Stats();
-			$stats       = $wpcom_stats->get_visits(
-				array(
-					'unit'        => 'day',
-					'quantity'    => 7,
-					'stat_fields' => 'visitors',
-				)
-			);
-
-			if ( ! is_wp_error( $stats ) && is_array( $stats ) ) {
-				return self::sum_visits_field( $stats, 'visitors' );
-			}
-		}
-
-		return 0;
-	}
-
-	/**
-	 * Fetch the last week's visitor count directly on WordPress.com Simple.
-	 */
-	private static function get_wpcom_simple_post_publish_promo_visitor_count(): int {
-		if ( ! function_exists( 'stats_get_visitors' ) ) {
-			return 0;
-		}
-
-		$visitors = stats_get_visitors( get_current_blog_id(), gmdate( 'Y-m-d' ), 7, 1 );
-
-		return is_array( $visitors ) ? (int) array_sum( $visitors ) : 0;
-	}
-
-	/**
-	 * Sum a metric from the Stats visits response.
-	 *
-	 * @param array  $stats Stats visits response.
-	 * @param string $field Field to sum.
-	 * @return int
-	 */
-	private static function sum_visits_field( array $stats, string $field ): int {
-		if ( ! isset( $stats['data'] ) || ! is_array( $stats['data'] ) ) {
-			return 0;
-		}
-
-		$fields = isset( $stats['fields'] ) && is_array( $stats['fields'] ) ? $stats['fields'] : array();
-		$index  = array_search( $field, $fields, true );
-		if ( false === $index ) {
-			return 0;
-		}
-
-		$total = 0;
-		foreach ( $stats['data'] as $row ) {
-			if ( is_array( $row ) && isset( $row[ $index ] ) ) {
-				$total += (int) $row[ $index ];
-			}
-		}
-
-		return $total;
-	}
-
-	/**
-	 * Whether the site is relevant for the post-publish promo.
-	 */
-	public static function is_post_publish_promo_site_eligible(): bool {
-		$host = new Host();
-		if ( $host->is_p2_site() ) {
-			return false;
-		}
-
-		if ( get_user_option( self::POST_PUBLISH_PROMO_DISMISSED_OPTION, get_current_user_id() ) ) {
-			return false;
-		}
-
-		return self::has_enough_recent_posts_for_post_publish_promo() && self::has_visitors_for_post_publish_promo();
 	}
 
 	/**
@@ -405,7 +180,6 @@ class Create_AI_Podcast_Page {
 				'loadingPosts'        => __( 'Loading posts…', 'jetpack-podcast' ),
 				'pickPosts'           => __( 'Select at least one post to continue.', 'jetpack-podcast' ),
 				'upgradeCta'          => __( 'Upgrade plan', 'jetpack-podcast' ),
-				'episodesTitle'       => __( 'Generated podcasts', 'jetpack-podcast' ),
 				'episodesEmpty'       => __( 'No generated podcasts yet.', 'jetpack-podcast' ),
 				'episodesLoading'     => __( 'Loading podcasts…', 'jetpack-podcast' ),
 				'editPost'            => __( 'Edit post', 'jetpack-podcast' ),
@@ -595,9 +369,7 @@ class Create_AI_Podcast_Page {
 	}
 
 	/**
-	 * Accept either the new envelope shape (preferred) or the legacy bare
-	 * array (older sandboxes that haven't shipped the pagination upgrade yet)
-	 * and return the envelope.
+	 * Normalize an upstream episodes response into the envelope shape.
 	 *
 	 * @param mixed $payload Upstream response body.
 	 *
@@ -611,16 +383,6 @@ class Create_AI_Podcast_Page {
 				'page'       => isset( $payload['page'] ) ? max( 1, (int) $payload['page'] ) : 1,
 				'perPage'    => isset( $payload['perPage'] ) ? max( 1, (int) $payload['perPage'] ) : self::EPISODES_PER_PAGE,
 				'totalPages' => isset( $payload['totalPages'] ) ? max( 0, (int) $payload['totalPages'] ) : 0,
-			);
-		}
-		if ( is_array( $payload ) ) {
-			$items = array_values( $payload );
-			return array(
-				'items'      => $items,
-				'total'      => count( $items ),
-				'page'       => 1,
-				'perPage'    => self::EPISODES_PER_PAGE,
-				'totalPages' => count( $items ) > 0 ? 1 : 0,
 			);
 		}
 		return self::empty_episodes_envelope();
@@ -654,10 +416,8 @@ class Create_AI_Podcast_Page {
 
 		$blog_id = (int) get_current_blog_id();
 
-		if ( function_exists( '\\Automattic\\Posts_To_Podcast\\get_usage' ) || function_exists( 'posts_to_podcast_get_usage' ) ) {
-			$usage              = function_exists( 'posts_to_podcast_get_usage' )
-				? posts_to_podcast_get_usage( $blog_id )
-				: array();
+		if ( function_exists( 'posts_to_podcast_get_usage' ) ) {
+			$usage              = posts_to_podcast_get_usage( $blog_id );
 			$bootstrap['quota'] = array(
 				'quota'      => $usage,
 				'activeJob'  => self::build_active_job_payload_wpcom( $blog_id ),
@@ -665,59 +425,7 @@ class Create_AI_Podcast_Page {
 			);
 		}
 
-		$per_page = self::EPISODES_PER_PAGE;
-		$query    = new \WP_Query(
-			array(
-				'post_type'              => 'post',
-				'post_status'            => array( 'draft', 'publish' ),
-				'posts_per_page'         => $per_page,
-				'paged'                  => 1,
-				'orderby'                => 'date',
-				'order'                  => 'DESC',
-				'update_post_term_cache' => false,
-				'meta_query'             => array(
-					array(
-						'key'     => 'posts_to_podcast_metadata',
-						'compare' => 'EXISTS',
-					),
-				),
-			)
-		);
-
-		$items = array();
-		foreach ( $query->posts as $post ) {
-			$raw_meta = get_post_meta( $post->ID, 'posts_to_podcast_metadata', true );
-			$meta     = is_string( $raw_meta ) ? json_decode( $raw_meta, true ) : null;
-			$audio    = ( is_array( $meta ) && isset( $meta['audio'] ) && is_array( $meta['audio'] ) ) ? $meta['audio'] : array();
-			$title    = wp_strip_all_tags(
-				html_entity_decode( (string) get_the_title( $post ), ENT_QUOTES | ENT_HTML5, 'UTF-8' )
-			);
-			if ( '' === trim( $title ) ) {
-				// translators: Fallback shown in the Generated podcasts list when a draft has an empty title.
-				$title = __( '(no title)', 'jetpack-podcast' );
-			}
-
-			$items[] = array(
-				'id'        => $post->ID,
-				'title'     => $title,
-				'status'    => $post->post_status,
-				'date'      => mysql2date( 'c', $post->post_date_gmt, false ),
-				'editUrl'   => get_edit_post_link( $post->ID, 'raw' ),
-				'mediaUrl'  => isset( $audio['url'] ) ? esc_url_raw( (string) $audio['url'] ) : '',
-				'mediaType' => 'audio',
-				'mediaMime' => isset( $audio['mimeType'] ) ? (string) $audio['mimeType'] : '',
-				'duration'  => isset( $audio['durationSeconds'] ) ? (int) round( (float) $audio['durationSeconds'] ) : 0,
-			);
-		}
-
-		$total                 = (int) $query->found_posts;
-		$bootstrap['episodes'] = array(
-			'items'      => $items,
-			'total'      => $total,
-			'page'       => 1,
-			'perPage'    => $per_page,
-			'totalPages' => (int) ceil( $total / $per_page ),
-		);
+		$bootstrap['episodes'] = Endpoint::build_episodes_envelope( 1, self::EPISODES_PER_PAGE );
 
 		return $bootstrap;
 	}
