+++ projects/packages/podcast/webpack.config.blocks.js
@@ -18,7 +18,7 @@ const sharedWebpackConfig = {
 		'podcast-episode/editor': './src/blocks/podcast-episode/editor.ts',
 		'podcast-episode/style': './src/blocks/podcast-episode/style.scss',
 		'podcast-episode/view': './src/blocks/podcast-episode/view.ts',
-		'post-publish-podcast-promo/editor': './src/editor/post-publish-podcast-promo/index.tsx',
+		'post-publish-podcast-promo/editor': './src/posts-to-podcast/editor-promo/index.tsx',
 	},
 	output: {
 		...jetpackWebpackConfig.output,
